Order new components in Login
