--- src/components/Home/2.header/2.1.Login/Login.jsx
@@ -1,6 +1,0 @@
-export default function Login(){
-  return(
-    <p>ESTO ES EL LOGIN</p>
-  );
-}
-
