+++ src/components/Home/2.login/b1.ModalLoginButtons.jsx
@@ -1,0 +1,13 @@
+export default function ModalLoginButtons(){
+  return(
+    <section id="switchButtons" class="bg-orange-200 flex items-center w-full h-[10%]">
+      <a href="#" class="w-1/2 h-full p-[1.8%] border-4 flex justify-center items-center bg-orange-500">      
+        <button >Ingresar</button>
+      </a>
+
+      <a href="#" class="w-1/2 h-full p-[1.8%] border-4 flex justify-center items-center">
+        <button>Quiero unirme!</button>
+      </a>
+    </section>
+  );
+}
